{{cookiecutter.out_dir}}/src/{{cookiecutter.plone_project_name}}/upgrades.py
# -*- coding: utf-8 -*-
from plone.app.upgrade.utils import loadMigrationProfile


def reload_gs_profile(context):
    loadMigrationProfile(
        context,
        'profile-{{cookiecutter.plone_project_name}}:default',
    )
